+++ apps/web/src/auth/useSignOut.ts
@@ -1,17 +1,15 @@
 import { clearTemporaryMasterKey } from "./AuthorizeActionDialog"
 import { SIGNAL_RESET } from "reactjrx"
-import { clearActiveProfileId, deleteProfileRow, getProfile } from "../profiles"
+import { clearActiveProfileId, getProfile, useDeleteProfile } from "../profiles"
 import { setUser } from "@sentry/react"
 import { googleAccessTokenSignal } from "../google/auth"
 import { usePluginsSignOut } from "../plugins/usePluginsSignOut"
-import { useQueryClient } from "@tanstack/react-query"
-import { profileByIdQueryKey } from "../profiles/useProfileById"
 import { useResetSessionQueries } from "../queries/useResetSessionQueries"
 
 export const useSignOut = () => {
   const signOutPlugins = usePluginsSignOut()
-  const queryClient = useQueryClient()
   const resetSessionQueries = useResetSessionQueries()
+  const { mutate: deleteProfile } = useDeleteProfile()
 
   return () => {
     const activeProfileId = getProfile()
@@ -24,10 +22,7 @@ export const useSignOut = () => {
     clearActiveProfileId()
 
     if (activeProfileId) {
-      queryClient.removeQueries({
-        queryKey: profileByIdQueryKey(activeProfileId),
-      })
-      void deleteProfileRow(activeProfileId)
+      deleteProfile(activeProfileId)
     }
 
     resetSessionQueries()
