+++ apps/web/src/profiles/dbHelpers.ts
@@ -2,5 +2,3 @@ import { dexieDb } from "../rxdb/dexie"
 import type { Profile } from "./types"
 
 export const putProfileRow = (row: Profile) => dexieDb.profiles.put(row)
-
-export const deleteProfileRow = (id: string) => dexieDb.profiles.delete(id)
