+++ apps/web/src/profiles/index.ts
@@ -28,3 +28,4 @@ export * from "./active/useIsActiveProfileHydrated"
 export * from "./active/useActiveProfile"
 export * from "./dbHelpers"
 export * from "./useProfileById"
+export * from "./useDeleteProfile"
